+++ otherfile/version_check.py
@@ -3,7 +3,7 @@
 import os
 import sys
 import codecs
-from art.art_param import *
+from art.params import *
 
 Failed = 0
 VERSION = "6.2"
@@ -40,7 +40,7 @@ FILES = {
     "ArtList.ipynb": ART_LIST_ITEMS,
     os.path.join(
         "art",
-        "art_param.py"): PARAMS_ITEMS,
+        "params.py"): PARAMS_ITEMS,
     os.path.join("otherfile", "meta.yaml"): META_ITEMS,
     os.path.join(".github", "ISSUE_TEMPLATE", "bug_report.yml"): ISSUE_TEMPLATE_ITEMS,
 }