+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `data` directory
+- `tests` directory
 - 10 new font
 	1. lolie
 	2. zakia
@@ -18,6 +20,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 	9. callian
 	10. awa
 ### Changed
+- `art_param` module changed to `params`
+- `art` module split into `errors`, `utils`, and `functions` modules
+- `text_dic1`, `text_dic2`, and `text_dic3` renamed to `fonts1`, `fonts2`, and `fonts3`
+- `art_dic` renamed to `arts`
+- `decor_dic` renamed to `decorations`
+- Fonts, arts, and decorations moved to `data` directory
+- Tests moved to `tests` directory
 - `Python 3.5` support dropped
 - Test system modified
 ## [6.2] - 2024-04-22